Uploaded source code of version 2.4.0 rc1.

+++ ayaya/commands/funny/Bigtext.java
@@ -1,15 +1,20 @@
 package ayaya.commands.funny;
 
 import ayaya.commands.Command;
+import ayaya.core.BotData;
 import ayaya.core.utils.SQLController;
 import com.jagrosh.jdautilities.command.CommandEvent;
 import net.dv8tion.jda.api.Permission;
+import net.dv8tion.jda.api.entities.Message;
 import net.dv8tion.jda.api.exceptions.ErrorResponseException;
 import net.dv8tion.jda.api.requests.ErrorResponse;
 import net.dv8tion.jda.api.requests.RestAction;
 
-import java.sql.*;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.util.Arrays;
 import java.util.EnumSet;
+import java.util.List;
 
 import static ayaya.core.enums.CommandCategories.FUNNY;
 
@@ -19,6 +24,9 @@ import static ayaya.core.enums.CommandCategories.FUNNY;
 public class Bigtext extends Command {
 
     private static final int LIMIT = 2000;
+    private static final List<Message.MentionType> ALLOWED_MENTIONS = Arrays.asList(
+            Message.MentionType.CHANNEL, Message.MentionType.EMOTE
+    );
 
     public Bigtext() {
 
@@ -40,16 +48,18 @@ public class Bigtext extends Command {
             event.reply("<:AyaWhat:362990028915474432> You want a big text of what? The air?");
         else {
             RestAction.setDefaultFailure(ErrorResponseException.ignore(EnumSet.of(ErrorResponse.UNKNOWN_MESSAGE)));
-            if (event.getChannelType().isGuild() && event.getGuild() != null &&
+            if (!event.getMessage().getMentions(
+                        Message.MentionType.USER, Message.MentionType.HERE, Message.MentionType.EVERYONE
+                ).isEmpty()
+                    && event.getChannelType().isGuild() && event.getGuild() != null &&
                     event.getSelfMember().hasPermission(event.getTextChannel(), Permission.MESSAGE_MANAGE))
                 event.getMessage().delete().queue();
-            String big_text = "From " + event.getAuthor().getAsTag()
-                    + " (" + event.getAuthor().getId() + "): " + getBigText(message);
+            String big_text = event.getAuthor().getAsMention() + ": " + getBigText(message);
             if (big_text.length() > LIMIT) {
                 event.replyError("The input message is too long. Remember that 1 emote is more than 1 character.");
                 return;
             }
-            event.reply(big_text);
+            event.getChannel().sendMessage(big_text).allowedMentions(ALLOWED_MENTIONS).queue();
         }
 
     }
@@ -66,17 +76,13 @@ public class Bigtext extends Command {
         SQLController jdbc = new SQLController();
         int i = 0;
         try {
-            jdbc.open("jdbc:sqlite:data.db");
+            jdbc.open(BotData.getDBConnection(), BotData.getDBUser(), BotData.getDbPassword());
             ResultSet result;
             do {
                 c = args.charAt(i);
-                result = jdbc.sqlSelect("SELECT * FROM `emojis` WHERE `emoji name` LIKE '" + c + "';", 5);
-                if (result.next()) {
-                    big_text.append(result.getString("emoji"));
-                } else {
-                    big_text.append(c);
-                }
-                big_text.append(" ");
+                result = jdbc.sqlSelect("SELECT * FROM emojis WHERE emoji_name='" + c + "';", 5);
+                big_text.append(result.next() ? result.getString("emoji") : c)
+                        .append(" ");
                 i++;
             } while (i < args.length());
         } catch (SQLException e) {
